transport: Move http functions to transport package
Allow external code to use HttpGet, HttpPatch and HttpPost functions.
