+++ internal/http_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/foundriesio/fioconfig/transport"
 	"github.com/stretchr/testify/require"
 )
 
@@ -18,7 +19,7 @@ func TestHttpRetry(t *testing.T) {
 
 	testWrapper(t, doGet, func(app *App, client *http.Client, tempdir string) {
 		start := time.Now()
-		res, err := httpGet(client, app.configUrl, nil)
+		res, err := transport.HttpGet(client, app.configUrl, nil)
 		elapsed := time.Since(start)
 		require.Nil(t, err)
 		require.Equal(t, 200, res.StatusCode)
@@ -28,7 +29,7 @@ func TestHttpRetry(t *testing.T) {
 		// Now do a test that won't retry:
 		codes = []int{400}
 		idx = 0
-		res, err = httpGet(client, app.configUrl, nil)
+		res, err = transport.HttpGet(client, app.configUrl, nil)
 		require.Nil(t, err)
 		require.Equal(t, 400, res.StatusCode)
 		require.Equal(t, 1, idx)
