+++ internal/rotate_config.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/ThalesIgnite/crypto11"
 	"github.com/foundriesio/fioconfig/sotatoml"
+	"github.com/foundriesio/fioconfig/transport"
 )
 
 type fullCfgStep struct{}
@@ -66,7 +67,7 @@ func (s deviceCfgStep) Execute(handler *certRotationContext) error {
 
 	// Download/decrypt current device config with current key
 	url := handler.app.configUrl + "-device"
-	res, err := httpGet(handler.client, url, nil)
+	res, err := transport.HttpGet(handler.client, url, nil)
 	if err != nil {
 		return err
 	}
@@ -110,7 +111,7 @@ func (s deviceCfgStep) Execute(handler *certRotationContext) error {
 			OnChanged:   entry.OnChanged,
 		})
 	}
-	res, err = httpPatch(handler.client, handler.app.configUrl, ccr)
+	res, err = transport.HttpPatch(handler.client, handler.app.configUrl, ccr)
 	if err != nil {
 		return err
 	}
